+++ nix/packages/nxr.nix
@@ -8,7 +8,7 @@
 
 rustPlatform.buildRustPackage {
   pname = "nxr";
-  version = "3.1.0";
+  version = "3.1.1";
 
   inherit src;